LayoutTests/crypto/subtle/rsa-pss-import-pkcs8-key-expected.txt

Test importing a PKCS8 RSA-PSS key

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a key...
PASS privateKey.toString() is '[object CryptoKey]'
PASS privateKey.type is 'private'
PASS privateKey.extractable is true
PASS privateKey.algorithm.name is 'RSA-PSS'
PASS privateKey.algorithm.modulusLength is 2048
PASS bytesToHexString(privateKey.algorithm.publicExponent) is '010001'
PASS privateKey.algorithm.hash.name is 'SHA-1'
PASS privateKey.usages is ['sign']
PASS successfullyParsed is true

TEST COMPLETE
